rename from behaviour to transform which describes better what it does
